Added collision systems. attempted multiple approaches - but ultimately settled on corner approach (checking intersections of 4 corners with tiles). This has some restrictions but good for now - need to change architecture of checking collisions but it seems okay for now.

+++ src/main/java/entities/constants/EntityConstants.java
@@ -4,5 +4,5 @@ package entities.constants;
 //into an Entities object.
 public abstract class EntityConstants implements AnimationConstants, HitboxConstants{
 
-    public abstract float getSpeed();
+    public abstract int getSpeed();
 }
